rikstoto-ai-wrapper/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic>=2.7.4
python-dotenv==1.0.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0